% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semibart.R
\name{semibart}
\alias{semibart}
\title{Semiparametric regression using BART}
\usage{
semibart(x.train, a.train, y.train, sigest = NA, sigdf = 3,
  sigquant = 0.9, k = 2, power = 2, base = 0.95, meanb = rep(0,
  ncol(a.train)), sigb = 4, ntree = 200, ndpost = 1000, numcut = 100,
  usequants = 1, offset = 0, binarylink = "probit", verbose = 1,
  printevery = 100)
}
\arguments{
\item{x.train}{Design matrix of values to be modeled with BART}

\item{a.train}{Design matrix of values to be modeled linearly}

\item{y.train}{Vector of outcome (continuous or binary)}

\item{sigest}{Estimate of regression error standard deviation}

\item{sigdf}{Prior value on error term}

\item{sigquant}{Prior value on error term}

\item{k}{Controls how aggressive the fit is}

\item{power}{Prior on tree depth}

\item{base}{Prior on tree depth}

\item{meanb}{Prior mean on regression coefficients -- length must equal # columns in a.train}

\item{sigb}{Prior standard deviation on regression coefficients ~ sigb^2 I}

\item{ntree}{Number of trees to use for BART}

\item{ndpost}{Number of MCMC iterations}

\item{numcut}{Number of cutpoints for each variable in BART. Must be of length 1 or have length ncol(x.train)}

\item{usequants}{Indicates whether to use observed quantiles for cutpoints or evenly spaced cutpoints based on min and max}

\item{offset}{Offset for regression -- used only when outcome is binary}

\item{binarylink}{Indicates whether to use probit or logit link for binary data}

\item{verbose}{Indicates whether or not user wants printed output}

\item{printevery}{Printevery}
}
\value{
Returns matrix of regression parameters (ndpost x ncol(a.train)). When y.train is continuous also returns vector of draws of regression parameter.
}
\description{
Semiparametric regression using BART
}

