\name{smmbart_rcpp-package}
\alias{smmbart_rcpp-package}
\alias{smmbart_rcpp}
\docType{package}
\title{
\packageTitle{smmbart_rcpp}
}
\description{
\packageDescription{smmbart_rcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{smmbart_rcpp}
\packageIndices{smmbart_rcpp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{smmbart_rcpp}

Maintainer: \packageMaintainer{smmbart_rcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
